Adds support for new function GetAccountInfo which retrieves your holdings for an exchange

+++ alphapointhttp.go
@@ -350,8 +350,8 @@ func (a *Alphapoint) GetExchangeAccountInfo() (ExchangeAccountInfo, error) {
 	for i := 0; i < len(account.Currencies); i++ {
 		var exchangeCurrency ExchangeAccountCurrencyInfo
 		exchangeCurrency.CurrencyName = account.Currencies[i].Name
-		exchangeCurrency.TotalValue = account.Currencies[i].Balance
-		exchangeCurrency.Hold = account.Currencies[i].Hold
+		exchangeCurrency.TotalValue = float64(account.Currencies[i].Balance)
+		exchangeCurrency.Hold = float64(account.Currencies[i].Hold)
 
 		response.Currencies = append(response.Currencies, exchangeCurrency)
 	}
